+++ Tests/nt.Client.Integration.Tests/Features/WebThree/GetAllOwnedTokens/FetchAllOwnedTokensTests.cs
@@ -65,7 +65,7 @@
 
                 // Token Balance
 
-                var BalanceContainer = await HttpClient.GetJsonAsync<BalanceOfSharedResponse>(BalanceOfSharedRequest.RouteFactory(token));
+                var BalanceContainer = await HttpClient.GetJsonAsync<BalanceOfSharedResponse>(BalanceOfSharedRequest.RouteFactory((int)token));
 
                 ownedToken.Balance = BalanceContainer.Balance;
